+++ core/app_services.go
@@ -17,6 +17,7 @@ package core
 import (
 	"core-building-block/core/model"
 	"core-building-block/driven/storage"
+	"slices"
 	"time"
 
 	"github.com/google/uuid"
@@ -155,12 +156,16 @@ func (app *application) serGetAccounts(limit int, offset int, appID string, orgI
 }
 
 func (app *application) serGetPublicAccounts(appID string, orgID string, limit int, offset int, search *string, firstName *string,
-	lastName *string, username *string, followingID *string, followerID *string, userID string) ([]model.PublicAccount, error) {
+	lastName *string, username *string, followingID *string, followerID *string, unstructuredProperties map[string]string, userID string) ([]model.PublicAccount, error) {
 	//find the accounts
-	accounts, err := app.storage.FindPublicAccounts(nil, appID, orgID, &limit, &offset, search, firstName, lastName, username, followingID, followerID, userID)
+	accounts, err := app.storage.FindPublicAccounts(nil, appID, orgID, &limit, &offset, search, firstName, lastName, username, followingID, followerID, unstructuredProperties, userID)
 	if err != nil {
 		return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeAccount, nil, err)
 	}
+
+	slices.SortFunc(accounts, func(i, j model.PublicAccount) int {
+		return i.OrderForSort(&j)
+	})
 
 	return accounts, nil
 }