+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Public account filtering by unstructured properties [#734](https://github.com/rokwire/core-building-block/issues/734)
+
 ## [1.44.1] - 2024-12-03
 ### Fixed
 - Fix Privacy.FieldVisibility validation [#729](https://github.com/rokwire/core-building-block/issues/729)